# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ActressTranslationView:
    properties:
      actress_id:
        type: integer
      created_at:
        type: string
      display_name:
        type: string
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      language:
        type: string
      last_name:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.ActressView:
    properties:
      aliases:
        type: string
      created_at:
        type: string
      dmm_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      name_key:
        type: string
      origin:
        type: string
      thumb_url:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressTranslationView'
        type: array
      updated_at:
        type: string
      verified:
        type: boolean
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult:
    properties:
      actress_sources:
        additionalProperties:
          type: string
        type: object
      ended_at:
        type: string
      error:
        type: string
      error_code:
        type: string
      field_sources:
        additionalProperties:
          type: string
        type: object
      file_path:
        type: string
      is_multi_part:
        type: boolean
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
      movie_id:
        type: string
      part_number:
        type: integer
      part_suffix:
        type: string
      result_id:
        description: Stable UUID — survives movie_id changes
        type: string
      revision:
        description: |-
          Revision: server-managed update counter used as the CAS baseline for
          PATCH expected_result_revision (POSTER-WRITE-HARDENING D12).
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus'
      translation_warning:
        description: |-
          TranslationWarning is populated when metadata translation partially failed
          or degraded; empty when translation fully succeeded.
        type: string
      translation_warning_code:
        description: |-
          TranslationWarningCode is the machine-readable classification of
          TranslationWarning (rate_limited, unavailable, degraded, ...).
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobListResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse'
        type: array
      total:
        description: Total number of jobs (before pagination), for pagination metadata
        type: integer
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse:
    properties:
      apply_generation:
        type: integer
      apply_plan:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.Plan'
      completed:
        type: integer
      completed_at:
        type: string
      destination:
        type: string
      excluded:
        additionalProperties:
          type: boolean
        type: object
      failed:
        type: integer
      files:
        items:
          type: string
        type: array
      id:
        type: string
      noop_count:
        description: |-
          NoopCount: terminal completed-noop rows (authorized duplicate skips —
          codex P2, PR #241 F2). Non-revertible; revertible = operation_count −
          reverted_count − noop_count.
        type: integer
      operation_count:
        type: integer
      operation_mode_override:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_operationmode.OperationMode'
      persist_error:
        type: string
      progress:
        type: number
      results:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult'
        type: object
      reverted_count:
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus'
      total_files:
        type: integer
      update:
        type: boolean
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchScrapeRequest:
    properties:
      apply_plan:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.Plan'
      array_strategy:
        description: 'For Update mode: merge, replace'
        example: merge
        type: string
      destination:
        description: Persisted on job for UI retrieval; required for organize mode,
          optional for in-place modes
        example: /path/to/output
        type: string
      files:
        items:
          type: string
        type: array
      force:
        example: false
        type: boolean
      manual_inputs:
        additionalProperties:
          type: string
        description: Per-file manual input override keyed by file path; an ID scrapes
          as that ID (bypasses matcher), a URL scrapes with URL-compatible scrapers
        example:
          '{"/path/file.mp4"': '"IPX-123"}'
        type: object
      operation_mode:
        description: 'Override config.output.operation_mode: organize, in-place, in-place-norenamefolder,
          metadata-artwork, preview'
        example: organize
        type: string
      preset:
        description: 'Merge strategy preset: conservative, gap-fill, aggressive (overrides
          scalar/array strategies)'
        example: conservative
        type: string
      scalar_strategy:
        description: 'For Update mode: prefer-nfo, prefer-scraper, preserve-existing,
          fill-missing-only'
        example: prefer-nfo
        type: string
      selected_scrapers:
        example:
        - r18dev
        - dmm
        items:
          type: string
        type: array
      strict:
        example: false
        type: boolean
      update:
        description: 'Update mode: only create/update metadata files without moving
          video files'
        example: false
        type: boolean
    required:
    - files
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchScrapeResponse:
    properties:
      job_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse:
    properties:
      code:
        description: Stable machine-readable error code (e.g. "JOB_NOT_FOUND") for
          client-side i18n
        type: string
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
      params:
        additionalProperties: {}
        description: Structured arguments for translating Code into a localized message
        type: object
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.GenreTranslationView:
    properties:
      created_at:
        type: string
      genre_id:
        type: integer
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.GenreView:
    properties:
      id:
        type: integer
      name:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreTranslationView'
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView:
    properties:
      created_at:
        type: string
      description:
        type: string
      director:
        type: string
      id:
        type: integer
      label:
        type: string
      language:
        type: string
      maker:
        type: string
      movie_id:
        type: string
      original_title:
        type: string
      series:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieView:
    properties:
      actresses:
        description: |-
          Relationships (contract DTOs — see collection_views.go; persistence
          models.* types never cross the API boundary)
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressView'
        type: array
      cast_version:
        type: string
      code:
        description: Canonical JAV code (e.g., "IPX-535"). Renamed from content_id.
        example: IPX-535
        type: string
      cover_url:
        type: string
      created_at:
        description: Audit timestamps
        type: string
      cropped_poster_url:
        type: string
      description:
        description: Metadata
        type: string
      director:
        type: string
      display_title:
        description: Titles
        example: IPX-535 Beautiful Woman
        type: string
      genres:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreView'
        type: array
      id:
        description: Secondary identifier — often equals Code but can differ.
        example: IPX-535
        type: string
      label:
        description: Sub-label
        type: string
      maker:
        description: Studio/maker
        type: string
      original_cover_url:
        type: string
      original_cropped_poster_url:
        type: string
      original_filename:
        type: string
      original_poster_url:
        type: string
      original_should_crop_poster:
        type: boolean
      original_title:
        example: 美人
        type: string
      poster_crop_bounds:
        allOf:
        - $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.CropBounds'
        description: Poster / cover (flattened from PosterState — no custom marshaler
          needed)
      poster_crop_source_full:
        type: boolean
      poster_url:
        type: string
      rating_score:
        type: number
      rating_votes:
        type: integer
      rating_warning:
        type: string
      release_date:
        type: string
      release_year:
        type: integer
      runtime:
        description: in minutes
        type: integer
      screenshot_urls:
        items:
          type: string
        type: array
      series:
        description: Series name
        type: string
      should_crop_poster:
        type: boolean
      source_name:
        description: Source provenance
        type: string
      source_url:
        type: string
      title:
        example: Beautiful Woman
        type: string
      trailer_url:
        description: Media
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_applyplan.ArrayStrategy:
    enum:
    - merge
    - replace
    type: string
    x-enum-varnames:
    - ArrayMerge
    - ArrayReplace
  github_com_javinizer_javinizer-go_internal_applyplan.MediaPolicy:
    enum:
    - missing
    - replace
    - skip
    type: string
    x-enum-varnames:
    - MediaPolicyMissing
    - MediaPolicyReplace
    - MediaPolicySkip
...
  github_com_javinizer_javinizer-go_internal_applyplan.MergePolicy:
    properties:
      array_strategy:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.ArrayStrategy'
      scalar_strategy:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.ScalarStrategy'
      source_preset:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.Preset'
    type: object
  github_com_javinizer_javinizer-go_internal_applyplan.NFOOutput:
    enum:
    - write
    - skip
    type: string
    x-enum-varnames:
    - NFOOutputWrite
    - NFOOutputSkip
  github_com_javinizer_javinizer-go_internal_applyplan.Plan:
    properties:
      destination:
        type: string
      media_policy:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.MediaPolicy'
      merge:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.MergePolicy'
      nfo_output:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.NFOOutput'
      version:
        type: integer
      video_operation:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.VideoOperation'
    type: object
  github_com_javinizer_javinizer-go_internal_applyplan.Preset:
    enum:
    - conservative
    - gap-fill
    - aggressive
    type: string
    x-enum-varnames:
    - PresetConservative
    - PresetGapFill
    - PresetAggressive
  github_com_javinizer_javinizer-go_internal_applyplan.ScalarStrategy:
    enum:
    - prefer-nfo
    - prefer-scraper
    - preserve-existing
    - fill-missing-only
    type: string
    x-enum-varnames:
    - ScalarPreferNFO
    - ScalarPreferScraper
    - ScalarPreserveExisting
    - ScalarFillMissingOnly
  github_com_javinizer_javinizer-go_internal_applyplan.VideoOperation:
    enum:
    - organize
    - rename-in-place
    - rename-file
    - leave-in-place
    - metadata-artwork
    type: string
    x-enum-varnames:
    - VideoOperationOrganize
    - VideoOperationRenameInPlace
    - VideoOperationRenameFile
    - VideoOperationLeaveInPlace
    - VideoOperationMetadataArtwork
...
  github_com_javinizer_javinizer-go_internal_models.CropBounds:
    properties:
      height:
        type: number
      source_aspect:
        description: |-
          SourceAspect is the width/height ratio of the source image the crop was
          measured against. The apply phase refuses geometry whose aspect no
          longer matches the downloaded image, so same-URL source swaps fall back
          to pre-change behavior instead of cropping the wrong image.
        type: number
      source_fingerprint:
        description: |-
          SourceFingerprint binds the normalized geometry to the exact full-size
          bytes measured in the review UI. Empty is retained for legacy envelopes;
          non-empty values must be a canonical SHA-256 hex digest.
        type: string
      width:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.JobStatus:
    enum:
    - pending
    - running
    - completed
    - failed
    - cancelled
    - organized
    - reverted
    type: string
    x-enum-varnames:
    - JobStatusPending
    - JobStatusRunning
    - JobStatusCompleted
    - JobStatusFailed
    - JobStatusCancelled
    - JobStatusOrganized
    - JobStatusReverted
...
  github_com_javinizer_javinizer-go_internal_operationmode.OperationMode:
    enum:
    - organize
    - in-place
    - in-place-norenamefolder
    - metadata-artwork
    - preview
    type: string
    x-enum-varnames:
    - OperationModeOrganize
    - OperationModeInPlace
    - OperationModeInPlaceNoRenameFolder
    - OperationModeMetadataArtwork
    - OperationModePreview
...
host: localhost:8765
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/batch:
    get:
      description: Get a paginated list of batch jobs with operation counts. Only
        job metadata is included; use GET /batch/{id} with include_data=true for full
        results.
      parameters:
      - default: 50
        description: Maximum number of jobs to return (default 50, max 200)
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of jobs to skip (default 0)
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: List batch jobs
      tags:
      - web
  /api/v1/batch/{id}:
    delete:
      description: Delete a completed or cancelled batch job and its temp files. Running
        jobs must be cancelled first.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "410":
          description: job already deleted (tombstone)
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Delete batch job
      tags:
      - web
    get:
      description: Retrieve the status of a batch scraping job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Get batch job status
      tags:
      - web
  /api/v1/batch/{id}/cancel:
    post:
      description: Cancel a running batch scraping job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Cancel batch job
      tags:
      - web
...
  /api/v1/batch/scrape:
    post:
      consumes:
      - application/json
      description: Scrape metadata for multiple movies in batch. Automatically discovers
        and includes all parts of multi-part files.
      parameters:
      - description: Batch scrape parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchScrapeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchScrapeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Batch scrape movies
      tags:
      - web
...
schemes:
- http
- https
swagger: "2.0"